+++ core/src/main/java/com/xatkit/core/recognition/DefaultFallbackIntent.java
@@ -1,0 +1,17 @@
+package com.xatkit.core.recognition;
+
+import com.xatkit.intent.impl.IntentDefinitionImpl;
+
+/**
+ * An {@link com.xatkit.intent.IntentDefinition} representing the default fallback.
+ */
+public class DefaultFallbackIntent extends IntentDefinitionImpl {
+
+    /**
+     * Constructs a new {@link DefaultFallbackIntent}.
+     */
+    DefaultFallbackIntent() {
+        super();
+        this.name = "Default_Fallback_Intent";
+    }
+}