+++ core/src/main/java/com/xatkit/core/recognition/IntentRecognitionProvider.java
@@ -1,100 +1,60 @@
 package com.xatkit.core.recognition;
 
-import com.xatkit.core.XatkitCore;
 import com.xatkit.core.recognition.processor.InputPreProcessor;
 import com.xatkit.core.recognition.processor.IntentPostProcessor;
 import com.xatkit.core.session.XatkitSession;
 import com.xatkit.intent.EntityDefinition;
 import com.xatkit.intent.IntentDefinition;
-import com.xatkit.intent.IntentFactory;
 import com.xatkit.intent.RecognizedIntent;
-import fr.inria.atlanmod.commons.log.Log;
 import lombok.NonNull;
-import org.apache.commons.configuration2.Configuration;
 
 import javax.annotation.Nullable;
-import java.util.ArrayList;
 import java.util.List;
 
 /**
- * A unified wrapper for concrete intent recognition providers.
+ * A unified wrapper to configure and query an intent recognition service.
  * <p>
- * This interface provides utility methods to connect to a given intent recognition provider, register and delete
- * {@link IntentDefinition}s, create {@link XatkitSession}s, and compute the {@link RecognizedIntent} from a given
- * input text.
+ * This interface allows to specify pre/post processors, register intents and entities to the intent recognition
+ * service, and retrieve {@link RecognizedIntent} instances from input text.
  */
-public abstract class IntentRecognitionProvider {
+public interface IntentRecognitionProvider {
 
     /**
-     * The Default Fallback Intent that is returned when the user input does not match any registered Intent.
+     * The Default fallback intent that is returned when the user input does not match any registered intent.
      */
-    public static IntentDefinition DEFAULT_FALLBACK_INTENT = IntentFactory.eINSTANCE.createIntentDefinition();
-
+    DefaultFallbackIntent DEFAULT_FALLBACK_INTENT = new DefaultFallbackIntent();
     /*
-     * Initializes the {@link #DEFAULT_FALLBACK_INTENT}'s name.
+     * We use a package private class here because we cannot statically initialize fields in interfaces (and we need
+     * to set the name of the IntentDefinition).
      */
-    static {
-        DEFAULT_FALLBACK_INTENT.setName("Default_Fallback_Intent");
-    }
-
-
-    /**
-     * The {@link List} of {@link InputPreProcessor}s set for this {@link IntentRecognitionProvider}.
-     *
-     * @see IntentRecognitionProviderFactory#getIntentRecognitionProvider(XatkitCore, Configuration)
-     * @see #getIntent(String, XatkitSession)
-     */
-    private List<? extends InputPreProcessor> preProcessors = new ArrayList<>();
-
-    /**
-     * The {@link List} of {@link IntentPostProcessor}s set for this {@link IntentRecognitionProvider}.
-     *
-     * @see IntentRecognitionProviderFactory#getIntentRecognitionProvider(XatkitCore, Configuration)
-     * @see #getIntent(String, XatkitSession)
-     */
-    private List<? extends IntentPostProcessor> postProcessors = new ArrayList<>();
 
     /**
      * Sets the {@link InputPreProcessor}s.
      *
      * @param preProcessors the {@link InputPreProcessor} to set
-     * @throws NullPointerException if the provided {@code preProcessors} is {@code null}
      */
-    public final void setPreProcessors(@NonNull List<? extends InputPreProcessor> preProcessors) {
-        this.preProcessors = preProcessors;
-    }
+    void setPreProcessors(@NonNull List<? extends InputPreProcessor> preProcessors);
+
+    /**
+     * Returns the {@link InputPreProcessor}s associated to this provider.
+     *
+     * @return the {@link InputPreProcessor}s associated to this provider
+     */
+    List<? extends InputPreProcessor> getPreProcessors();
 
     /**
      * Set the {@link IntentPostProcessor}s.
      *
      * @param postProcessors the {@link IntentPostProcessor} to set
-     * @throws NullPointerException if the provided {@code postProcessors} is {@code null}
      */
-    public final void setPostProcessors(@NonNull List<? extends IntentPostProcessor> postProcessors) {
-        this.postProcessors = postProcessors;
-    }
+    void setPostProcessors(@NonNull List<? extends IntentPostProcessor> postProcessors);
 
     /**
-     * Returns the {@link InputPreProcessor}s associated to this {@link IntentRecognitionProvider}.
-     * <p>
-     * This method is package-private for testing purposes.
+     * Returns the {@link IntentPostProcessor}s associated to this provider.
      *
-     * @return the {@link InputPreProcessor}s associated to this {@link IntentRecognitionProvider}
+     * @return the {@link IntentPostProcessor}s associated to this provider
      */
-    final List<? extends InputPreProcessor> getPreProcessors() {
-        return this.preProcessors;
-    }
-
-    /**
-     * Returns the {@link IntentPostProcessor}s associated to this {@link IntentRecognitionProvider}.
-     * <p>
-     * This method is package-private for testing purposes.
-     *
-     * @return the {@link IntentPostProcessor}s associated to this {@link IntentRecognitionProvider}
-     */
-    final List<? extends IntentPostProcessor> getPostProcessors() {
-        return this.postProcessors;
-    }
+    List<? extends IntentPostProcessor> getPostProcessors();
 
     /**
      * Registers the provided {@code entityDefinition} in the underlying intent recognition provider.
@@ -105,10 +65,9 @@ public abstract class IntentRecognitionProvider {
      * {@link #trainMLEngine()} to train the ML engine.
      *
      * @param entityDefinition the {@link EntityDefinition} to delete from the underlying intent recognition provider
-     * @throws NullPointerException if the provided {@code entityDefinition} is {@code null}
      * @see #trainMLEngine()
      */
-    public abstract void registerEntityDefinition(@NonNull EntityDefinition entityDefinition);
+    void registerEntityDefinition(@NonNull EntityDefinition entityDefinition);
 
     /**
      * Registers the provided {@code intentDefinition} in the underlying intent recognition provider.
@@ -119,10 +78,9 @@ public abstract class IntentRecognitionProvider {
      * {@link #trainMLEngine()} to train the ML engine.
      *
      * @param intentDefinition the {@link IntentDefinition} to register to the underlying intent recognition provider
-     * @throws NullPointerException if the provided {@code intentDefinition} is {@code null}
      * @see #trainMLEngine()
      */
-    public abstract void registerIntentDefinition(@NonNull IntentDefinition intentDefinition);
+    void registerIntentDefinition(@NonNull IntentDefinition intentDefinition);
 
     /**
      * Deletes the provided {@code entityDefinition} from the underlying intent recognition provider.
@@ -133,10 +91,9 @@ public abstract class IntentRecognitionProvider {
      * {@link #trainMLEngine()} to train the ML engine.
      *
      * @param entityDefinition the {@link EntityDefinition} to delete from the underlying intent recognition provider
-     * @throws NullPointerException if the provided {@code entityDefinition} is {@code null}
      * @see #trainMLEngine()
      */
-    public abstract void deleteEntityDefinition(@NonNull EntityDefinition entityDefinition);
+    void deleteEntityDefinition(@NonNull EntityDefinition entityDefinition);
 
     /**
      * Deletes the provided {@code intentDefinition} from the underlying intent recognition provider.
@@ -147,10 +104,9 @@ public abstract class IntentRecognitionProvider {
      * {@link #trainMLEngine()} to train the ML engine.
      *
      * @param intentDefinition the {@link IntentDefinition} to delete from the underlying intent recognition provider
-     * @throws NullPointerException if the provided {@code intentDefinition} is {@code null}
      * @see #trainMLEngine()
      */
-    public abstract void deleteIntentDefinition(@NonNull IntentDefinition intentDefinition);
+    void deleteIntentDefinition(@NonNull IntentDefinition intentDefinition);
 
     /**
      * Trains the underlying intent recognition provider.
@@ -158,16 +114,15 @@ public abstract class IntentRecognitionProvider {
      * <b>Note:</b> this method returns once the intent recognition provider's training is complete. However, the
      * propagation of the training information may not be complete when this method returns.
      */
-    public abstract void trainMLEngine();
+    void trainMLEngine();
 
     /**
      * Creates a new {@link XatkitSession} from the provided {@code sessionId}.
      *
      * @param sessionId the identifier to create a session from
      * @return a new {@link XatkitSession} for the provided {@code sessionId}
-     * @throws NullPointerException if the provided {@code sessionId} is {@code null}
      */
-    public abstract XatkitSession createSession(@NonNull String sessionId);
+    XatkitSession createSession(@NonNull String sessionId);
 
     /**
      * Shuts down the intent recognition provider client and invalidates the remaining sessions.
@@ -175,14 +130,14 @@ public abstract class IntentRecognitionProvider {
      * <b>Note:</b> calling this method invalidates the intent recognition provider client connection, and thus this
      * class cannot be used to access the intent recognition provider anymore.
      */
-    public abstract void shutdown();
+    void shutdown();
 
     /**
      * Returns whether the intent recognition provider client is shutdown.
      *
      * @return {@code true} if the intent recognition provider client is shutdown, {@code false} otherwise
      */
-    public abstract boolean isShutdown();
+    boolean isShutdown();
 
     /**
      * Returns the {@link RecognizedIntent} extracted from te provided {@code input}.
@@ -190,59 +145,25 @@ public abstract class IntentRecognitionProvider {
      * This method uses the provided {@code session} to extract contextual intents, such as follow-up or
      * context-based intents.
      * <p>
-     * This method applies the pre-processing functions associated to this {@link IntentRecognitionProvider} on the
-     * given {@code input}, and the post-processing functions on the returned {@link RecognizedIntent}.
+     * This method applies the pre-processing functions associated to this provider on the given {@code input}, and
+     * the post-processing functions on the returned {@link RecognizedIntent}.
+     * <p>
+     * This method returns a {@link RecognizedIntent} with {@code RecognizedIntent#getDefinition() ==
+     * DEFAULT_FALLBACK_INTENT} if the provided {@code input} does not match any intent.
      *
      * @param input   the {@link String} representing the textual input to process and extract the intent from
      * @param session the {@link XatkitSession} used to access context information
      * @return the post-processed {@link RecognizedIntent} extracted from the provided {@code input} and {@code session}
-     * @throws NullPointerException if the provided {@code input} or {@code session} is {@code null}
-     * @see IntentRecognitionProviderFactory#getIntentRecognitionProvider(XatkitCore, Configuration)
+     * @see #DEFAULT_FALLBACK_INTENT
      */
-    public final RecognizedIntent getIntent(@NonNull String input, @NonNull XatkitSession session) {
-        String preProcessedInput = input;
-        for (InputPreProcessor preProcessor : this.preProcessors) {
-            long preStart = System.currentTimeMillis();
-            preProcessedInput = preProcessor.process(input, session);
-            long preEnd = System.currentTimeMillis();
-            Log.debug("Time to execute pre-processor {0}: {1}ms", preProcessor.getClass().getSimpleName(),
-                    (preEnd - preStart));
-        }
-        long recognitionStart = System.currentTimeMillis();
-        RecognizedIntent recognizedIntent = getIntentInternal(preProcessedInput, session);
-        long recognitionEnd = System.currentTimeMillis();
-        Log.debug("Time to recognize the intent with {0}: {1}ms", this.getClass().getSimpleName(),
-                (recognitionEnd - recognitionStart));
-        for (IntentPostProcessor postProcessor : this.postProcessors) {
-            long postStart = System.currentTimeMillis();
-            recognizedIntent = postProcessor.process(recognizedIntent, session);
-            long postEnd = System.currentTimeMillis();
-            Log.debug("Time to execute post-processor {0}: {1}ms", postProcessor.getClass().getSimpleName(),
-                    (postEnd - postStart));
-        }
-        return recognizedIntent;
-    }
-
-    /**
-     * Returns the raw {@link RecognizedIntent} extracted from the provided {@code input}.
-     * <p>
-     * This method is called <b>after</b> pre-processing of the {@code input} (i.e. the given {@code input} is
-     * already pre-processed), and does not apply any post-processing function (this is done by
-     * {@link #getIntent(String, XatkitSession)}). Subclasses implementing this method should not take care of
-     * pre/post processing.
-     *
-     * @param input   the textual input to process and extract the intent from
-     * @param session the {@link XatkitSession} used to access context information
-     * @return the {@link RecognizedIntent} extracted from the provided {@code input} and {@code session}
-     * @throws NullPointerException if the provided {@code input} or {@code session} is {@code null}
-     */
-    protected abstract RecognizedIntent getIntentInternal(@NonNull String input, @NonNull XatkitSession session);
+    @NonNull RecognizedIntent getIntent(@NonNull String input, @NonNull XatkitSession session);
 
     /**
      * Returns the {@link RecognitionMonitor} associated to this intent recognition provider.
      *
      * @return the {@link RecognitionMonitor}, or {@code null} if analytics monitoring is disabled.
      */
-    public abstract @Nullable
+    @Nullable
     RecognitionMonitor getRecognitionMonitor();
+
 }